+++ src/bundler/rollupPlugins/resolveOtherBundles.js
@@ -1,25 +1,28 @@
-import { transformAsync } from '@babel/core';
-import * as t from '@babel/types';
+import MagicString from 'magic-string';
 import { dirname, relative as getRelativePath, resolve as resolvePath } from 'pathe';
+import resolveDTSImportPaths from '../resolveDTSImportPaths.js';
 
-export default function resolveOtherBundles( projectPath, metadata ) {
+export default function resolveOtherBundles( projectPath, metadata, {
+	isTypeBundling = false
+} = {} ) {
 	return {
 		name: 'rlb-resolve-other-bundles',
 
 		async resolveId( importee, importer ) {
-			if ( !importee.startsWith( '.' ) ) {
+			if ( !importee.startsWith( '.' ) || !importer ) {
 				return null;
 			}
 
-			const resolved = await this.resolve( importee, importer, {
-				skipSelf: true
-			} );
-			const srcFullPath = resolved.id;
-			const srcPathRelativeToProject = getRelativePath( projectPath, srcFullPath );
-			const isBundle = typeof metadata[ srcPathRelativeToProject ] !== 'undefined';
+			const resolved = isTypeBundling ?
+				resolveDTSImportPaths( importee, importer ) :
+				await this.resolve( importee, importer, {
+					skipSelf: true
+				} );
+			const srcPathRelativeToProject = getSrcPathRelativeToProject( projectPath, resolved, isTypeBundling );
+			const isBundle = checkIfBundle( srcPathRelativeToProject, metadata, isTypeBundling );
 
 			if ( !isBundle ) {
-				return null;
+				return isTypeBundling ? resolved.id : null;
 			}
 
 			const distPlaceholderPath = `rlb:${ srcPathRelativeToProject }`;
@@ -32,11 +35,16 @@ export default function resolveOtherBundles( projectPath, metadata ) {
 
 		async renderChunk( code, chunk, { file } ) {
 			const chunkFullPath = resolvePath( projectPath, file );
-			const { code: transformedCode, map } = await transformAsync( code, {
-				plugins: [
-					transformImports( projectPath, metadata, chunkFullPath )
-				]
+			const magicString = new MagicString( code );
+
+			magicString.replaceAll( /(?:import|export).+?from\s*["'](rlb:.+?)["']/g, ( importOrExportString, importee ) => {
+				const bundlePath = getCorrectImportPath( importee, metadata, chunkFullPath );
+
+				return importOrExportString.replace( importee, bundlePath );
 			} );
+
+			const transformedCode = magicString.toString();
+			const map = magicString.generateMap();
 
 			return {
 				code: transformedCode,
@@ -44,68 +52,8 @@ export default function resolveOtherBundles( projectPath, metadata ) {
 			};
 		}
 	};
-}
 
-function transformImports( projectPath, metadata, importerFullPath ) {
-	return {
-		visitor: {
-			ImportDeclaration( path ) {
-				const { node } = path;
-				const { value: importee } = node.source;
-
-				if ( !importee.startsWith( 'rlb:' ) ) {
-					return;
-				}
-
-				const importPath = getCorrectImportPath( importee );
-
-				path.replaceWith(
-					t.importDeclaration(
-						node.specifiers,
-						t.stringLiteral( importPath )
-					)
-				);
-			},
-
-			ExportNamedDeclaration( path ) {
-				const { node } = path;
-				const { source } = node;
-
-				if ( !source || !source.value.startsWith( 'rlb:' ) ) {
-					return;
-				}
-
-				const importPath = getCorrectImportPath( source.value );
-
-				path.replaceWith(
-					t.exportNamedDeclaration(
-						node.declaration,
-						node.specifiers,
-						t.stringLiteral( importPath )
-					)
-				);
-			},
-
-			ExportAllDeclaration( path ) {
-				const { node } = path;
-				const { source } = node;
-
-				if ( !source.value.startsWith( 'rlb:' ) ) {
-					return;
-				}
-
-				const importPath = getCorrectImportPath( source.value );
-
-				path.replaceWith(
-					t.exportAllDeclaration(
-						t.stringLiteral( importPath )
-					)
-				);
-			}
-		}
-	};
-
-	function getCorrectImportPath( importee ) {
+	function getCorrectImportPath( importee, metadata, importerFullPath ) {
 		const srcPathRelativeToProject = importee.slice( 4 );
 		const distPathRelativeToProject = metadata[ srcPathRelativeToProject ].esm;
 		const distFullPath = resolvePath( projectPath, distPathRelativeToProject );
@@ -116,5 +64,23 @@ function transformImports( projectPath, metadata, importerFullPath ) {
 			`./${ distPathRelativeToChunk }`;
 
 		return importPath;
+	}
+
+	function getSrcPathRelativeToProject( projectPath, resolved, isTypeBundling ) {
+		if ( isTypeBundling ) {
+			return resolved.tsSourceFilePath;
+		}
+
+		return getRelativePath( projectPath, resolved.id );
+	}
+
+	function checkIfBundle( srcPath, metadata, isTypeBundling ) {
+		const isEntryPreset = typeof metadata[ srcPath ] !== 'undefined';
+
+		if ( !isTypeBundling ) {
+			return isEntryPreset;
+		}
+
+		return isEntryPreset && metadata[ srcPath ].types;
 	}
 }